WsdlWar needed some safety catches in the constructor.  It gets evaluated initially, even though it doesn't do anything with it (apparently), but when it gets run for real it goes through it again.
Running without the check if the dependencies were null was
croaking the execution of the task, even without running the dependent
tasks!

Also saw that when there were no xsd dependencies it brought in
the whole folder -- this is because if there were no dependencies, there
would be no include patters, and the default on a tree with no includes
or excludes is EVERYTHING under that tree.

A simple check before the relPaths are calculated is to see if the
dependencies are empty and if they are return an empty fileCollection
via project.files()